Enhancements based on code review, upgrade to .NET 4.5, other enhancements.
Upgrade to .NET 4.5 and VS 2012.

Rename AngelList to AngelList.Interface

"In JsonNetAngelListClient, There's code to do pagination and handling
1st page. " - method to call  first page calls method to handle
pagination with page =1

Removed "Not Implemented" code.

"AngelList.JsonTypes has 2 Location classes": re-generated Json types so
that there are no duplicate classes and all are in the same namespace,
which is much simpler to use and read.

Added .gitignore file.

In database, added Investor.Updated and Industry.Updated to allow
querying for data that needs to be updated.